requirements.txt
paho-mqtt==2.1.0
pymongo==4.15.1
requests==2.32.5
pytest==8.4.2
flask==3.1.0
python-dotenv==1.0.1